+++ examples/lzapp-migration/tasks/index.ts
@@ -1,8 +1,8 @@
 import './common/config.get'
 import './common/taskHelper'
 import './common/wire'
-import './sendV1Message'
-import './sendV2Message'
+import './evm/sendV1Message'
+import './evm/sendV2Message'
 
 import './evm/epV1oftv2Send'
